requirements.txt
asgiref==3.6.0
cffi==1.15.1
cfgv==3.3.1
cryptography==39.0.2
distlib==0.3.6
Django==4.1.7

django-rest-knox==4.2.0
djangorestframework==3.14.0

filelock==3.9.0
identify==2.5.19
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==3.1.0
pre-commit==3.1.1
pycparser==2.21
PyJWT==1.7.1
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
virtualenv==20.20.0
